@@ -1,6 +1,7 @@
 {
   "name": "nuxt-vite-tailwindcss",
   "version": "1.0.0",
+  "author": "Vuong Phan <info@codefix.xyz>",
   "private": true,
   "scripts": {
     "dev": "nuxt",
@@ -23,16 +24,18 @@
     }
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.12.5",
-    "core-js": "^3.8.3",
-    "nuxt": "^2.14.12"
+    "@nuxtjs/axios": "^5.13.1",
+    "core-js": "^3.9.1",
+    "nuxt": "^2.15.3",
+    "nuxt-i18n": "^6.21.1"
   },
   "devDependencies": {
     "@commitlint/cli": "^12.0.1",
     "@commitlint/config-conventional": "^12.0.1",
     "@nuxt/image": "^0.4.1",
-    "@nuxt/types": "^2.14.12",
-    "@nuxt/typescript-build": "^2.0.4",
+    "@nuxt/types": "^2.15.3",
+    "@nuxt/typescript-build": "^2.1.0",
+    "@nuxtjs/composition-api": "0.22.1",
     "@nuxtjs/eslint-config-typescript": "^6.0.0",
     "@nuxtjs/eslint-module": "^3.0.2",
     "@nuxtjs/google-fonts": "^1.3.0",
@@ -42,26 +45,27 @@
     "@tailwindcss/forms": "^0.2.1",
     "@tailwindcss/line-clamp": "^0.2.0",
     "@tailwindcss/typography": "^0.4.0",
-    "@vue/test-utils": "^1.1.2",
+    "@vue/composition-api": "1.0.0-rc.5",
+    "@vue/test-utils": "^1.1.3",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.3",
-    "eslint": "^7.18.0",
+    "eslint": "^7.22.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-nuxt": "^2.0.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^7.5.0",
+    "eslint-plugin-vue": "^7.7.0",
     "husky": "^5.1.3",
     "jest": "^26.6.3",
-    "lint-staged": "^10.5.3",
+    "lint-staged": "^10.5.4",
     "nuxt-vite": "^0.0.36",
     "postcss": "^8.2.8",
     "prettier": "^2.2.1",
-    "stylelint": "^13.9.0",
+    "stylelint": "^13.12.0",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-standard": "^21.0.0",
     "tailwindcss": "^2.0.4",
     "ts-jest": "^26.5.4",
-    "vue-jest": "^3.0.4"
+    "vue-jest": "^3.0.7"
   }
 }
