@@ -1,5 +1,5 @@
-import { mount } from '@vue/test-utils'
 import Logo from '@/components/Logo.vue'
+import { mount } from '@vue/test-utils'
 
 describe('Logo', () => {
   test('is a Vue instance', () => {
